+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ Changelog
 
 **Improvements**:
 
+- ``dir(resource)`` and ``list(resource)`` now also show properties of an object
 - *Pro Edition:* Added support for RedmineUP DealCategory ``create()``, ``update()``, ``delete()`` operations
   (see `docs <https://python-redmine.com/resources/deal_category.html#create-methods>`__ for details)
 - *Pro Edition:* RedmineUP CrmQuery resource now supports ``invoices`` and ``expenses`` relation attributes